Accept String as package name & move ReflectionUtils to config package.
