+++ hazelcast-jet-core/src/test/kotlin/com/hazelcast/jet/config/KotlinReflectionUtilsTest.kt
@@ -1,24 +1,24 @@
-package com.hazelcast.jet
+package com.hazelcast.jet.config
 
 import org.hamcrest.Matchers.containsInAnyOrder
 import org.hamcrest.Matchers.hasSize
 import org.junit.Assert.assertThat
 import org.junit.Test
 
-class KotlinReflectionsTest {
+class KotlinReflectionUtilsTest {
 
     @Test
     fun shouldAddNestedAndAnonymousClasses() {
         // When
-        val classes = Reflections.memberClassesOf(OuterClass::class.java)
+        val classes = ReflectionUtils.memberClassesOf(OuterClass::class.java)
 
         // Then
         assertThat(classes, hasSize(4))
         assertThat(classes, containsInAnyOrder(
                 OuterClass::class.java,
                 OuterClass.NestedClass::class.java,
-                Class.forName("com.hazelcast.jet.KotlinReflectionsTest\$OuterClass\$method\$1"),
-                Class.forName("com.hazelcast.jet.KotlinReflectionsTest\$OuterClass\$method\$lambda\$1")
+                Class.forName("com.hazelcast.jet.config.KotlinReflectionUtilsTest\$OuterClass\$method\$1"),
+                Class.forName("com.hazelcast.jet.config.KotlinReflectionUtilsTest\$OuterClass\$method\$lambda\$1")
         ))
     }
